refactor: fetch cards in SetPage, and pass card info and onClick handler into CardList
